Add success messages, which I'll call `Reactions`
This implementation assumes there's just one best reaction. But our
content inventory contains more, so maybe these will later be a
random selection, for replay value.
